fix: remove REPLACE MySQL specific functionality for #39
