--- test/phpunit/Helper/Query/ReplaceInferredPlaceholderExample.php
@@ -1,19 +1,0 @@
-<?php
-namespace Gt\SqlBuilder\Test\Helper\Query;
-
-use Gt\SqlBuilder\Query\ReplaceQuery;
-
-class ReplaceInferredPlaceholderExample extends ReplaceQuery {
-	public function into():array {
-		return [
-			"student",
-		];
-	}
-
-	public function set():array {
-		return [
-			"name",
-			"dateOfBirth",
-		];
-	}
-}
