--- test/phpunit/Helper/Query/ReplaceMixedPlaceholderExample.php
@@ -1,22 +1,0 @@
-<?php
-namespace Gt\SqlBuilder\Test\Helper\Query;
-
-use Gt\SqlBuilder\Query\ReplaceQuery;
-
-class ReplaceMixedPlaceholderExample extends ReplaceQuery {
-	public function into():array {
-		return [
-			"student",
-		];
-	}
-
-	public function set():array {
-		return [
-			"name",
-			"dateOfBirth",
-			"createdAt" => ":dateTimeNow",
-			"enabled" => 1,
-			"type",
-		];
-	}
-}
